[ 🔧 | #52 | refactor ] SchoolPage 리팩토링 완료
