+++ src/pages/SchoolPage/components/SchoolRankItem.jsx
