+++ src/pages/SchoolPage/components/SchoolRankList.jsx
@@ -1,5 +1,5 @@
-import SchoolRankItem from '@/pages/SchoolPage/SchoolRankList/SchoolRankItem';
-import { schoolRankData } from '@/pages/SchoolPage/SchoolRankList/_data/mock';
+import SchoolRankItem from '@/pages/SchoolPage/components/SchoolRankItem';
+import { schoolRankData } from '@/pages/SchoolPage/data/mock';
 
 export default function SchoolRankList() {
 	return (
